chore: add slots to keyboardcontrols
