+++ src/core/controls/index.ts
@@ -1,5 +1,5 @@
 import OrbitControls from './OrbitControls.vue'
-import { KeyboardControls } from './KeyboardControls'
+import KeyboardControls from './KeyboardControls.vue'
 import TransformControls from './TransformControls.vue'
 import PointerLockControls from './PointerLockControls.vue'
 import MapControls from './MapControls.vue'
